app/requirements.txt
torch>=2.2.0
gradio==3.40.0
wordcloud==1.8.1
matplotlib==3.5.1
transformers==4.32.1
sentencepiece==0.1.99
beautifulsoup4==4.11.1
indic-nlp-library==0.81
Flask==2.2.5
gTTS==2.3.2
vaderSentiment==3.3.2
